Revised with main-tests
Updated own .cpp and .h files to handle text code provided in main-tests branch main.cpp.

+++ add_subtract.h
@@ -14,7 +14,7 @@
 
 int improperFractionNumerator (int c1, int n1, int d1);
 int commonDenom (int d1, int d2);
-int moveDecimal(long numerator, int digitsLeft, int commonDenom);
+long long moveDecimal(long numerator, int digitsLeft, int commonDenom);
 bool add(int c1, int n1, int d1, int c2, int n2, int d2, char result[], int len);
 bool subtract(int c1, int n1, int d1, int c2, int n2, int d2, char result[], int len);
